计算金额超大

+++ src/main/java/com/aiqin/mgs/order/api/domain/response/CostAndSalesResp.java
@@ -23,11 +23,11 @@ public class CostAndSalesResp {
 
     @ApiModelProperty(value = "销售商品数量")
     @JsonProperty("sale_product_num")
-    private Integer saleProductNum;
+    private Long saleProductNum;
 
     @ApiModelProperty(value = "退货商品数量")
     @JsonProperty("return_product_num")
-    private Integer returnProductNum;
+    private Long returnProductNum;
 
     @ApiModelProperty(value = "销售金额")
     @JsonProperty("total_amount")
@@ -49,7 +49,7 @@ public class CostAndSalesResp {
 
     @ApiModelProperty("商品成本")
     @JsonProperty("product_cost")
-    private Integer productCost;
+    private Long productCost;
 
     @ApiModelProperty(value = "毛利额")
     @JsonProperty("product_maori")
